Absorb iris.x into Context, delete iris/x/
Move tile-level collective primitives (all_reduce_*, all_gather, gather,
all_to_all, reduce_scatter) from standalone iris.x functions into Context
methods. Move types (Tile, TileView, TensorView, AllReduceConfig) and
utilities into iris/mem/triton/types.py. Update all consumers in iris/ops/
and tests. Rename tests/x/ to tests/context/.
