+++ iris/ops/all_gather_matmul.py
@@ -14,7 +14,6 @@ import torch
 import triton
 import triton.language as tl
 import iris
-import iris.x
 from iris.host.tracing.kernel_artifacts import iris_launch
 
 from tritonblas.kernels.stages import GemmContext, ScheduleContext
@@ -83,7 +82,7 @@ def _fused_all_gather_matmul_kernel(
 
         # Create DeviceContext and TensorView for gather operations
         ctx = iris.DeviceContext.initialize(context_tensor, cur_rank, world_size)
-        src_view = iris.x.make_tensor_view(A_sharded, M, K_local, stride_am, stride_ak)
+        src_view = iris.make_tensor_view(A_sharded, M, K_local, stride_am, stride_ak)
 
         # Precompute B column offsets for this output tile (constant across K iterations)
         rn = pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)
@@ -102,10 +101,10 @@ def _fused_all_gather_matmul_kernel(
                 # Create tile view for this K block
                 # Promote tile_k to tensor (TileView expects tl.tensor for pid_n)
                 tile_k = pid_m * 0 + k_offset // BLOCK_SIZE_K
-                k_tile = iris.x.TileView(pid_m, tile_k, BLOCK_SIZE_M, BLOCK_SIZE_K)
+                k_tile = iris.TileView(pid_m, tile_k, BLOCK_SIZE_M, BLOCK_SIZE_K)
 
                 # Pull A tile from source_rank_id using gather primitive
-                a = iris.x.gather(k_tile, src_view, source_rank_id, ctx)
+                a = ctx.gather(k_tile, src_view, source_rank_id)
 
                 # Load B tile using direct pointer arithmetic
                 # Compute global K row index for B matrix
@@ -126,10 +125,10 @@ def _fused_all_gather_matmul_kernel(
                 k_offset = loop_k_local * BLOCK_SIZE_K
                 # Promote tile_k to tensor (TileView expects tl.tensor for pid_n)
                 tile_k = pid_m * 0 + k_offset // BLOCK_SIZE_K
-                k_tile = iris.x.TileView(pid_m, tile_k, BLOCK_SIZE_M, BLOCK_SIZE_K)
+                k_tile = iris.TileView(pid_m, tile_k, BLOCK_SIZE_M, BLOCK_SIZE_K)
 
                 # Pull A tile from source_rank_id using gather primitive
-                a = iris.x.gather(k_tile, src_view, source_rank_id, ctx)
+                a = ctx.gather(k_tile, src_view, source_rank_id)
 
                 # Load B tile with boundary handling
                 global_k_offset = source_rank_id * K_local + loop_k_local * BLOCK_SIZE_K
