+++ iris/ops/matmul_all_gather.py
@@ -15,7 +15,6 @@ import torch
 import triton
 import triton.language as tl
 import iris
-import iris.x
 from iris.host.tracing.kernel_artifacts import iris_launch
 
 from tritonblas.kernels.stages import GemmContext, ScheduleContext, make_tensor_view
@@ -97,12 +96,12 @@ def _fused_matmul_all_gather_kernel(
 
         # Create DeviceContext and destination TensorView for all-gather
         ctx = iris.DeviceContext.initialize(context_tensor, cur_rank, world_size)
-        dst_view = iris.x.make_tensor_view(C_gathered, M, N, stride_cm_gathered, stride_cn_gathered)
-        tile_obj = iris.x.Tile(out_tile.pid_m, out_tile.pid_n, BLOCK_SIZE_M, BLOCK_SIZE_N, c)
+        dst_view = iris.make_tensor_view(C_gathered, M, N, stride_cm_gathered, stride_cn_gathered)
+        tile_obj = iris.Tile(out_tile.pid_m, out_tile.pid_n, BLOCK_SIZE_M, BLOCK_SIZE_N, c)
 
-        # Scatter this tile to all ranks using iris.x.all_gather
+        # Scatter this tile to all ranks using all_gather
         # dim=0 means scatter along M dimension (rows)
-        iris.x.all_gather(tile_obj, dst_view, dim=0, ctx=ctx)
+        ctx.all_gather(tile_obj, dst_view, dim=0)
 
 
 def matmul_all_gather_preamble(
