+++ tests/context/__init__.py
@@ -2,5 +2,5 @@
 # Copyright (c) 2025 Advanced Micro Devices, Inc. All rights reserved.
 
 """
-Test suite for iris.x tile-level primitives.
+Test suite for Context tile-level collective methods.
 """
