+++ tests/context/test_all_to_all.py
@@ -11,7 +11,6 @@ import torch.distributed as dist
 import triton
 import triton.language as tl
 import iris
-import iris.x
 
 
 @triton.jit
@@ -43,12 +42,12 @@ def x_all_to_all_kernel(
         pid_n = tile_id % num_pid_n
 
         # Create OOP objects for new API
-        tile = iris.x.TileView(pid_m, pid_n, BLOCK_SIZE_M, BLOCK_SIZE_N)
-        src_view = iris.x.make_tensor_view(input_ptr, M, N, stride_in_m, stride_in_n)  # N is total N
-        dst_view = iris.x.make_tensor_view(output_ptr, M, N, stride_out_m, stride_out_n)  # N is total N
+        tile = iris.TileView(pid_m, pid_n, BLOCK_SIZE_M, BLOCK_SIZE_N)
+        src_view = iris.make_tensor_view(input_ptr, M, N, stride_in_m, stride_in_n)  # N is total N
+        dst_view = iris.make_tensor_view(output_ptr, M, N, stride_out_m, stride_out_n)  # N is total N
         ctx = iris.DeviceContext.initialize(context_tensor, cur_rank, world_size)
 
-        iris.x.all_to_all(tile, src_view, dst_view, N_per_rank, ctx)
+        ctx.all_to_all(tile, src_view, dst_view, N_per_rank)
 
 
 @pytest.mark.parametrize(
