+++ tests/context/test_reduce_scatter.py
@@ -11,7 +11,6 @@ import torch.distributed as dist
 import triton
 import triton.language as tl
 import iris
-import iris.x
 
 
 @triton.jit
@@ -56,12 +55,12 @@ def x_reduce_scatter_kernel(
         tl.debug_barrier()
         tl.atomic_xchg(locks + tile_id, 1, sem="release", scope="gpu")
 
-        tile = iris.x.Tile(pid_m, pid_n, BLOCK_SIZE_M, BLOCK_SIZE_N, local_data)
-        src_view = iris.x.make_tensor_view(temp_buffer, M, N, stride_in_m, stride_in_n)
-        dst_view = iris.x.make_tensor_view(output_ptr, M, N, stride_out_m, stride_out_n)
+        tile = iris.Tile(pid_m, pid_n, BLOCK_SIZE_M, BLOCK_SIZE_N, local_data)
+        src_view = iris.make_tensor_view(temp_buffer, M, N, stride_in_m, stride_in_n)
+        dst_view = iris.make_tensor_view(output_ptr, M, N, stride_out_m, stride_out_n)
         ctx = iris.DeviceContext.initialize(context_tensor, cur_rank, world_size)
 
-        iris.x.reduce_scatter(tile, src_view, dst_view, locks, ctx)
+        ctx.reduce_scatter(tile, src_view, dst_view, locks)
 
 
 @pytest.mark.parametrize(
